requirements.txt
fastapi
uvicorn[standard]
chromadb
openai
pydantic
python-dotenv
